* Complete program: This is a solution to the "Bot saves princess" challenge for Python 3 on hackerrank.com. The original problem can be found here:
https://www.hackerrank.com/challenges/saveprincess

The goal is to determine and print the shortest set of moves a hero character must travel from the center of a grid of size m to the corner of the grid in which the princess is located. The moves are printed one by one after the displayPathToPrincess function determines the hero's relative distance on X and Y from the princess.